lnd: extract neutrino initialization to new function
